rename hello to app
